# PERF: Scope performance.