Script used:
```ruby
begin
  require 'bundler/inline'
rescue LoadError => e
  $stderr.puts 'Bundler version 1.10 or later is required. Please update your Bundler'
  raise e
end

gemfile(true) do
  source 'https://rubygems.org'
  gem 'rails', path: '~/rails' # master against ref "f1f0a3f8d99aef8aacfa81ceac3880dcac03ca06"
  gem 'arel', github: 'rails/arel', branch: 'master'
  gem 'rack', github: 'rack/rack', branch: 'master'
  gem 'sass'
  gem 'sprockets-rails', github: 'rails/sprockets-rails', branch: 'master'
  gem 'sprockets', github: 'rails/sprockets', branch: 'master'
  gem 'pg'
  gem 'benchmark-ips'
end

require 'active_record'
require 'benchmark/ips'

ActiveRecord::Base.establish_connection('postgres://postgres@localhost:5432/rubybench')

ActiveRecord::Migration.verbose = false

ActiveRecord::Schema.define do
  create_table :users, force: true do |t|
    t.string :name, :email
    t.timestamps null: false
  end
end

class User < ActiveRecord::Base; end

attributes = {
  name: "Lorem ipsum dolor sit amet, consectetur adipiscing elit.",
  email: "foobar@email.com",
}

1000.times { User.create!(attributes) }

Benchmark.ips(5, 3) do |x|
  x.report('where with hash') { User.where(name: "Lorem ipsum dolor sit amet, consectetur adipiscing elit.") }
  x.report('where with string') { User.where("users.name = ?", "Lorem ipsum dolor sit amet, consectetur adipiscing elit.") }
  x.compare!
end

key =
  if RUBY_VERSION < '2.2'
    :total_allocated_object
  else
    :total_allocated_objects
  end

before = GC.stat[key]
User.where(name: "Lorem ipsum dolor sit amet, consectetur adipiscing elit.")
after = GC.stat[key]
puts "Total Allocated Object: #{after - before}"

Benchmark.ips(5, 3) do |x|
  x.report('all') { User.all }
end
```

Before:
```ruby
Calculating -------------------------------------
     where with hash     2.727k i/100ms
   where with string     4.202k i/100ms
-------------------------------------------------
     where with hash     28.893k (± 1.8%) i/s -    144.531k
   where with string     44.463k (± 3.0%) i/s -    222.706k

Comparison:
   where with string:    44463.1 i/s
     where with hash:    28893.2 i/s - 1.54x slower

Total Allocated Object: 85
Calculating -------------------------------------
                 all    16.155k i/100ms
-------------------------------------------------
                 all    177.896k (± 3.7%) i/s -    904.680k
```

After:
```ruby
Calculating -------------------------------------
     where with hash     3.207k i/100ms
   where with string     4.732k i/100ms
-------------------------------------------------
     where with hash     33.849k (± 1.1%) i/s -    169.971k
   where with string     49.735k (± 2.3%) i/s -    250.796k

Comparison:
   where with string:    49734.8 i/s
     where with hash:    33848.8 i/s - 1.47x slower

Total Allocated Object: 83
Calculating -------------------------------------
                 all    19.699k i/100ms
-------------------------------------------------
                 all    222.530k (± 2.6%) i/s -      1.123M
```